+++ Core/Src/Utils.c
@@ -7,9 +7,10 @@
 
 
 #include "Utils.h"
+
 #include "stm32f4xx_hal.h"
 
-uint16_t Driver_GetTimeMs(void)
+uint32_t Driver_GetTimeMs(void)
 {
-    return (uint16_t)HAL_GetTick();
+    return HAL_GetTick();
 }
